look for certificate.pem and private.key

+++ smoketest.bash
@@ -220,10 +220,15 @@ fi
 createProxyCertsVolume() {
     "${container_engine}" volume create auth_proxy_certs
     "${container_engine}" container create --name proxy_certs_helper -v auth_proxy_certs:/certs busybox
-    chmod 444 "${DIR}/compose/auth_certs/private.key"
-    chmod 444 "${DIR}/compose/auth_certs/certificate.pem"
-    "${container_engine}" cp "${DIR}/compose/auth_certs/certificate.pem" proxy_certs_helper:/certs/certificate.pem
-    "${container_engine}" cp "${DIR}/compose/auth_certs/private.key" proxy_certs_helper:/certs/private.key
+    if [ -f "${DIR}/compose/auth_certs/certificate.pem" ] && [ -f "${DIR}/compose/auth_certs/private.key" ]; then
+        chmod 444 "${DIR}/compose/auth_certs/private.key"
+        chmod 444 "${DIR}/compose/auth_certs/certificate.pem"
+        "${container_engine}" cp "${DIR}/compose/auth_certs/certificate.pem" proxy_certs_helper:/certs/certificate.pem
+        "${container_engine}" cp "${DIR}/compose/auth_certs/private.key" proxy_certs_helper:/certs/private.key
+    else
+        echo "Unable to find a certificate and key to allow oauth2_proxy to enable TLS connections"
+        exit 2
+    fi
 }
 if [ "${USE_PROXY}" = "true" ]; then
     createProxyCertsVolume
